Erk. Forwardport all of r875819. Thanks dfaure + pusling for noticing I haven't done this:
SVN commit 875819 by orlovich:
Fix how  we find existing frames for the purpose of form target= submission.
This fixes file upload for drupal 6/jquery.form --- its submit  was incorrectly
blocked as a popup attempt

As we refactored this method, also use it for window.open, fixing us blocking 
opens to _self along the way.

While testing this, I noticed that we can crash konq if we do window.open 
and one of the frames doesn't yet have a part yet. Added a workaround 
suggested by dfaure for now, as the proper fix is too risky and should be 
done on unstable branch only.

svn path=/trunk/KDE/kdelibs/; revision=877526