Set color before drawing line numbers.
Before, the folding markers would set the pen color and fail to restore it, causing the line numbers to be the wrong color (and IME often illegible). For now, set the pen before drawing line numbers. We could do better by properly combining set/restore cycles where the pen changes (this would lead to fewer overall sets), but we aren't very consistent about it now, so that's saved for a later cleanup (and because of that, remove the initial set since it is now superfluous).

BUG: 169419

svn path=/trunk/KDE/kdelibs/; revision=898209