fixed: no more endless accelerator update loop with tabbed QDockWidgets
The tab for a QDockWidget does always get synchronized to the title of the
QDockWidget. So setting different synchronizers is impossible, both will have the
same shortcut in the end -> conflict -> new accelerator manager run.

Solved by taking QTabBars for QDockWidgets out of management. The resulting same
shortcut for both tab and QDockWidget can be ignored, as a QDockWidget does not
grab a shortcut anyway. But giving no shortcut to a QDockWidget will also leave the
tab without a shortcut, due to the synchronisation. So picked the less(?) worse solution.

Still needed: better heuristic to detect a QTabBar for QDockWidgets.

See http://lists.kde.org/?l=kde-core-devel&m=123069857419153 and http://lists.kde.org/?l=kde-core-devel&m=123099108014556

svn path=/trunk/KDE/kdelibs/; revision=905350