backport commit 908324 from trunk
Bug in KDE with graphicssystem raster engine. To draw the itemViewPanelItem we use a QImage with the format ARGB32_Premultiplied but we fill it incorrectly.
QImage.fill take a uint not a QColor so this color was invalid. The result was an ugly yellow painting with the raster engine. I just still ask myself why x11 paint engine doesn't complain and was working with an invalid color.

Btw : This one was hard to catch!!!

svn path=/branches/KDE/4.2/kdelibs/; revision=908329