backport SVN commit 908320 by thiago:
Accept the -graphicssystem option even if we run with Qt 4.5, not
build against it.

This code will break in Qt 4.10, so I left an #error if we get that
far. Hopefully we'll have a function that returns the number instead
of the string. (I have that as a patch already)

svn path=/branches/KDE/4.2/kdelibs/; revision=908789