backport SVN commit 925654 by tilladam:
When using the QFileSystemWatcher backend of KDirWatch, trust it to correctly
notify us when a file or diretory changes, instead of second guessing it based
on mtime. Adding a file to a directory does not always update the directory
mtime, for example, which makes this check fail in some situations.

svn path=/branches/KDE/4.2/kdelibs/; revision=925692