Fix ruby syntax bugs
A single-quoted single-character string inside a shell command caused
the rest of the file to be incorrectly highlighted.

A general delimited input expression passed as the first argument to a
method without using parenthesis was incorrectly interpreted as modulo.

This bug was very common when using the puts method:
puts %(hello world)

svn path=/trunk/KDE/kdelibs/; revision=935483