Merged revision:r939356 | vtokarev | 2009-03-14 13:48:22 -0400 (Sat, 14 Mar 2009) | 6 lines
Avoid usage of QVector as temporary buffers. .clear() on it triggers
internal buffer deletion and shrinking capacity to 0. Which is not good
if you want to append something to it afterwards.
Replacing with saner WTF::Vector and QVarLengthArray, which gives significant
improvement for ProspectiveTokenizer. And less allocations too.
don't ever use QVector if you want to clear it a lot

svn path=/branches/KDE/4.2/kdelibs/; revision=942370