I overhauled the look of the File Transfer Dialog, shown when the plasma one is disabled. Yes, plasma is the default for transfers (but I like dialogs better for this, I actually think they should eventually make it like gnome had (I think), where you can click the icon in the tray to minimize the opened dialog to tray, that kind of smoothness should be added I think).
Anyways... This transfer dialog can be seen when disabling the plasma one (via systray settings)....

The dialog is now collapsed (by default), and can be expanded/toggled that way, by the arrow button.

Collapsed means that technical stats are hidden, things like transfer speed, time estimation, etc.

This is a step towards a cleaner transfer dialog, there could be more to go, I will look into it further......

Note that the green arrow is not what we should use (plasma also uses it for this), waiting for Oxygen team to make one (already filed a request), so this is only temporary.

P.S. No, I didn't take the idea from the plasma blog post, I actually started this before that, with the help of this pretty bug-report-mockup  ;-)

Here are screenies of what it is like in action here: 
1.) http://is.gd/pkM1
2.) http://is.gd/pkLY

Compared to before:
- http://is.gd/pkVA

A small step closer towards my goal of overhauling kdelibs' GUI nasties :P

BUG:160055

svn path=/trunk/KDE/kdelibs/; revision=945701