/Enormous/ speedup for rendering speed of large/text-heavy documents - takes /half/ the time as before this on the test files I have - e.g. some French dictionary file from Spart now loads in 12 seconds, while it used to take 24 (and 26 in 4.2.1); rendering of MySQL reference doc html takes 8 seconds and not 16.
Note: results are likely to be less significant on other sorts of workloads.

This change works by caching width information for fonts in a sparse 2-level
tree, avoiding /insanely/ expensive costs of doing that via Qt (testing
was done with Qt4.5, optimized build). Also a slight shrinkage to
InheritedStyleData, but probably not enough to matter.


svn path=/branches/KDE/4.2/kdelibs/; revision=946162