replace the kdeglobals handling with something slightly less bizarre
KConfig("kdeglobals") would have essentially opened the file as a both
local and global file. this required lots of special casing and was
still broken.
so now just degrade it to a normal file with no globals. there is still
some special handling due to the ominous system.kdeglobals support, but
it is within reason.

CCMAIL: faure@kde.org

svn path=/trunk/KDE/kdelibs/; revision=946222