Really new password strength computation
Well, my last commit fixed the 12345 issue, but I was not
really satisfied. With this new algorithm you will really
have a hard time to satisfy 100% because it computes the
number of character category changes, and changes between
vowels and consonants are NOT counted, because this will
look like a word!


svn path=/trunk/KDE/kdelibs/; revision=952601