I love it when something that takes days to track down ends up being a stupid, tiny brain fart like this. Let me tell you a little story ...
Sometimes, weirdly, Kontact's configuration dialog would hang on Windows, for
no apparent reason. After quite a bit of digging and trying to catch the hung
process in the debugger it turns out that it hangs while trying to access a
dbus interface, but only the second time you start the app. The reason for that
is that on the first start, the check whether the interface exists fails, and
thus registering it works. The second time, that check hangs. Weird. Well, why
is that interface broken? Until you access it, it seems the other interfaces
are fine, after that access they are all broken. Even weirder. Actually, even
if you remove the second interface, bringing up the config dialog screws all
dbus interfaces. Curious. Strangely, even with all of KCMultiDialog (the base
class) commented out, that still happens. What the hell does Kontact's config
dialog do then, to trigger this? Well, it calls KWindowSystem::setIcons, quite
innocently. That happens to be the first thing to call KWindowSystem::init(),
which calls KWindowSystem::activate(), which does ... nothing that seems out of
the ordinary, let alone access dbus. Curiouser and curiouser. Well, let's just
bisect that method until it starts working, and behold! Due to a kDebug()
statement right after an if without a curly brace, the logic is:

if ( do it one way )
  kDebug() << ....
  actually do it one way
  if ( something unrelated )
     do something unrelated
else {
  do it it the other way
  ...

So both ways are actually executed, which means a system hook which was
resolved dynamiclly from a dll is called twice, in different ways, which it
apparently does not like, mucking up some internal state of the process,
thereby rendering all the process' dbus interfaces unusable. Why exactly that
symptom happens is beyond me. This is clearly the cause, though. Bugger all, I
think I need a drink now. :)

Fixes kolab/issue3532.

P.S.: David, I think this is one for your talk at Akademy ;)
P.P.S.: Patrick, since you apparently wrote this, you owe me a beer ;))

svn path=/branches/kdepim/enterprise4/kdelibs-4.2-branch/; revision=966931