Fix crash in Wallpapers GHNS when no providers are found
The problem was this: If you have an Internet connection, you
can download the list of provided themes, and GHNS caches that.

If you later open that dialog, but have no Internet connection,
then the dialog will show cached entries, but still say "No
providers found", and abort the dialog. After that abort, the
engine is deleted, but the dialog is not closed.

When we later get a paint() event, we crash because the model
for the items has already been deleted.

The fix changes the slot to close the dialog before returning
to the event loop in case the providers were not found, so that
the dialog has a chance to delete the widgets before the engine
is deleted.

BUG: 167473
CCMAIL: jeremy@scitools.com

svn path=/trunk/KDE/kdelibs/; revision=972805