Fix the Y0K bug
The Julian/Gregorian Calendar does not have a Year 0, so all maths on a year
needs to adjust for this when moveing over the BC/AD divide.  A new private
method addYearNumber() does this correctly and should be used in all parts
of the code that needs to add years.

The Hebrew, Hijri and Jalali calendars are unaffected by this as none of our
implementations are proleptic, i.e. allow dates before year 1.

In 4.4 when we get calendar systems that do have a Year 0 we will need to
modify the addYearNumber() method to repect this.

Leave bug open for now as will need to backport to 4.3 once released.

CCBUG: 154016


svn path=/trunk/KDE/kdelibs/; revision=992292