merge r927883 | osterfeld | 2009-02-18 08:47:29 -0500 (Wed, 18 Feb 2009) | 7 lines
Wallet::openWallet() uses an event loop to wait for a dbus reply from kwalletd.
However, if the daemon crashes/is killed (as sometimes necessary on OS X) and never replies,
the app is stuck in the forever running event loop.
Thus keep a pointer to the event loop and quit the loop in case the daemon's dbus service goes away.


svn path=/branches/kdepim/enterprise4/kdelibs-4.2-branch/; revision=995807