KLocale::formatByteSize() improvements:
* Add the ability to specify a particular binary unit standard to
  KLocale::formatByteSize.
* Also add the ability to specify a particular binary unit (i.e. KiB, MiB) for
  situations where the auto-unit-sizing is not what is needed.
* If the default 1 digit of precision is not accurate enough, you can specify
  more precision.
* Dialects supported: IEC 60027-2 (-ibibytes), JEDEC (which are the traditional
  units standardized by JEDEC for use with semiconductor memory/storage sizes),
  and metric (kB, MB using standard SI prefixes instead of JEDEC).

The default binary unit dialect remains IEC as it is unambiguous and in
accordance with the most relevant standards bodies.  Because I love getting
emails about what is apparently a very political topic I've added the ability
to change the default using a hidden option, BinaryUnitDialect.  Set to 0 for
IEC, 1 for JEDEC, 2 for metric.  If you set it metric, you will receive metric
units (yes, really decimal powers of ten).  You have been warned, but if it's
what you've wanted all your life, KDE 4.4 will be the platform for you. ;)

There will be no GUI in the foreseeable future to change the unit type, just
use kate.

This has been discussed quite a bit on kde-devel and kde-core-devel, without as
much consensus as I'd like (but more than I'd hoped for).  In general, the idea
seems to be that if you really feel you need specific binary units, you
probably still don't.  Be careful about changing the unit type you display to
the user without a good reason.

Still TODO: a formatByteRate() function, and ensuring that the rest of KDE is
already using formatByteSize.  Based on kde-core-devel feedback I do not
intend to start changing any applications to display a specific unit type at
this point, but instead to simply prefer user default.

And oh yeah, testcases are included. :)

svn path=/trunk/KDE/kdelibs/; revision=996857