Distinguish between v4l1 and v4l2 devices. Patch reviewed on reviewboard.
This patch allows applications using solid for video device detection to distinguish between
video4linux1 and video4linux2 devices. This is useful to know because v4l1 and v4l2 use
different API, so the application needs to know which API to use to communicate with the device.
This is accomplished by returning "video4linux" for v4l1 devices and "video4linux2" for v4l2
devices as a supported driver for the "video4linux" protocol. The actual information is
retrieved from HAL, using the "video4linux.version" device property field.

svn path=/trunk/KDE/kdelibs/; revision=1006441