Fix the Y0K bug.  Backport commit 992292 from trunk.
The Julian/Gregorian Calendar does not have a Year 0, so all maths on a year
needs to adjust for this when moving over the BC/AD divide.  A new private
method addYearNumber() does this correctly and should be used in all parts
of the code that needs to add years.

The Hebrew, Hijri and Jalali calendars are unaffected by this as none of our
implementations are proleptic, i.e. allow dates before year 1.

In 4.4 when we get calendar systems that do have a Year 0 we will need to
modify the addYearNumber() method to repect this.

BUG: 154016


svn path=/branches/KDE/4.3/kdelibs/; revision=1007713