Fix sanitization of dbus path in KMainWindow.
As written, the illegal characters '.' and '-' would be allowed. This fixes it by only allowing letters and digits; everything else is replaced by '_' (also a legal character). We could test also for '_', but since it is replaced by itself, it would only be an optimization, and this spot shouldn't be so performance-critical that we need to worry about that.

Thanks go to Thomas Lübking for pointing out the problem, and for the patch.

REVIEW: 1261

svn path=/trunk/KDE/kdelibs/; revision=1012430