Fix sanitization of dbus path in KMainWindow, take 2.
The whole path - i.e. including appName() - is now sanitized, and to ASCII letters (as opposed to QChar::isLetter, which would also allow non-ASCII letters). The validation function is copied from qdbusutil.cpp (with the slight change that '/' is also allowed, since it works on the whole path, not just components).

Thanks go to Thiago for suggesting this approach.

REVIEW: 1353

svn path=/trunk/KDE/kdelibs/; revision=1013091