Mimetype determination: Fix glob matching according to discussion on xdg list, so that *.C, *.c and core are strictly case-sensitive-only.
This fixes the case of "foo.PS.gz", which the previous algorithm would treat as "simply gzip",
and it makes the matching of mixed-case extensions like "textfile.TxT" 3 times faster (QBENCHMARK+callgrind rocks).

This commit does -not- require the next release of shared-mime-info; we emulate what it will
output i.e. force those three hardcoded globs to case-sensitive if shared-mime-info
is <= 0.60 (which is the case for everyone right now), to avoid regressions.

BUG: 198477

svn path=/trunk/KDE/kdelibs/; revision=1013836