Fix KRestrictedLine to not let composed characters like éèú etc. go through, if they're not wanted.
And add test code for KRestrictedLine in klineedittest.
But with possible event compression in keyPressEvent and the way inputMethodEvent works, we really
get QStrings as input, not "int key" like the API assumes here.
So we should really use a validator instead, so we can remove unwanted chars and
keep the others, which means a different signal signature.
Dario, can you commit your validator, commented out, for kde5? :-)
CCBUG: 161012

svn path=/trunk/KDE/kdelibs/; revision=1014077