Fix crashes on exit of the command line runner for some modules.
This changes the unloading of a module to no longer be done
at library unload (which is when the singleton destructor is called),
but at exit of main.
This solves deletion order being different based on which context kross (the lib)
is used.  Now its explicit that module data (like documents) are deleted
first before any library unloading is done.

CCMAIL:koffice-devel@kde.org

svn path=/trunk/KDE/kdelibs/; revision=1027291