Prevent double-dispatch of events file upload lineedit does not accept. (One directly to lineedit, one via DOM on the containing upload widget).
Fixes double-upload-dialogs-on-enter bug. I think in general the bubbling 
behavior of Qt when it comes to events is problematic --- we should 
probably always accept events at original level and route them manually
to avoid issues. (This requires moving top-level shortcut handling 
from KHTMLView to ElementImpl::defaultEventHandler with a check 
for being the documentElement)

BUG: 192755


svn path=/trunk/KDE/kdelibs/; revision=1029831