Revert r477175:
When we unbind a smart range which has a dynamic attribute with animation set,
it's range has to be removed directly from the dynamicRegion(). Relying on the
animation to finish in time is not an option: It takes 300msec and in this time
the range could have been removed and would lead to dangling pointers
in KateSmartRegion.

Alternatively one could potentially use KateSmartRangePtr in KateSmartRegion.
If this is the preferred way, tell me and I'll change the code accordingly.

CCMAIL: rodda@kde.org

svn path=/trunk/KDE/kdelibs/; revision=1031007