change the way global settings are propagated
instead of KGlobalSettings' c'tor propagating them automatically,
introduce an activate() method which must be called explicitly (usually
by KApplication's c'tor).

this undoes gof's change to skip applying the settings if qApp is no
KApplication - it is unnecessary now.

while this is an incompatible change of (undocumented) behavior, the
collateral damage expected from it is zero.

this is necessary to fix a bug in kdm (currently the most frequently
reported one), so committing to the stable branch.

CCMAIL: faure@kde.org, ogoffart@kde.org


svn path=/branches/KDE/4.3/kdelibs/; revision=1033582