change the way global settings are propagated
instead of KGlobalSettings' c'tor propagating them automatically,
introduce an activate() method which must be called explicitly (usually
by KApplication's c'tor).

this undoes gof's change to skip applying the settings if qApp is no
KApplication - it is unnecessary now.

while this is an incompatible change of (undocumented) behavior, the
collateral damage expected from it is zero.

svn path=/trunk/KDE/kdelibs/; revision=1033885