Printing enhancemnt to add Odd/Even page selection and Server Side page selection via CUPS only.
Posted on http://reviewboard.kde.org/r/1785/ no objections received to the
slightly hacky nature of this change.  If issues arise please contact me
for alternative implementation.

With the printing enhancements not making it into Qt 4.6, we need an interim
fix to provide Odd/Even page selection and support of server side page range
selection.  This is possible by extending the current CUPS support to provide
these features under CUPS only.

1) Odd/Even pages.  This is achieved through a hack to add extra widgets to
the actual Qt dialog Copies tab by using findWidget() and insertWidget().

2) Server side page range selection.  Some apps like Kate and Konqi can't/don't
do page selection and force the user to print the entire document.  Cups does
support selecting the pages for the app and this change allows the app to set
a flag to tell it to do so.

Note also that I have now made all the KDE-added CUPS features only be added
if the Qt version is < 4.7 as I'm fairly certain the new features will be
available in Qt by then.


svn path=/trunk/KDE/kdelibs/; revision=1035322