Unbreaking pulse animation. It will handle the case where there is no 'shadow' widget (and thus not set the widget itself as its parent).
By connecting a slot to the signal emited by animation end, it will
also restore original opacity back.

TODO: instead of a copy of object, the shadow should be an image
where the animated widget was painted. This will solve a lot
of issues and make this class easier to use.


svn path=/trunk/KDE/kdelibs/; revision=1036351