Add Julian, Gregorian Proleptic, and Indian National Calendar Systems
To support the Indian National Calendar System (the official government if not
widely used calendar system in India) two new features were required.

1) Support for Year 0 in KCalendarSystem. All the currently supported Calendar
   systems do not have a Year 0 and the date math routines compensated for
   this.  This needed to be changed to allow both counting systems to be used.

2) A proleptic Gregorian calendar. The Indian National calendar is tightly
   linked to the Gregorian calculations, but the current 'Gregorian' class
   matches QDate in being a hybrid of Julian before 15 October 1582 and
   Gregorian afterwards, and so would produce incorrect results.

This proleptic Gregorian calendar will not be offered to users as a choice of
Calendar System as it would likely confuse them to have two Gregorians to
choose from, and it will produce inconsistant results with QDate and/or
Gregorian.  Coders who know what they are doing are welcome to use it.  Note
that this proleptic Gregorian chooses not to have a Year 0, unlike the ISO
version which does.

The Indian National calendar is a civil calendar, not one of the many Indian
religious calendars that require astronomical calculation support to be
implemented.  Hopefully these will follow in KDE 4.5.

And for good measure, I've added the Julian Calendar System as well which has
usage for a number of religions.

A overhauled Calendar System test suite will follow soon.


svn path=/trunk/KDE/kdelibs/; revision=1038093