First step to improve new animations memory model.
Now the QAbstractAnimations will be cached when running an animation
(instead of creating and destroying them everytime the animation runs).

The next step is to cache *also* the configuration parameters (as aseigo suggested,
by having an event filter to detect changes in the animated widgets and reseting
the animation parameters as needed).


svn path=/trunk/KDE/kdelibs/; revision=1039531