Add support for Coptic and Ethiopian calendar systems.
The Ethiopian calendar is the official calendar in Ethiopia, and is used in a
number of religions.

The Coptic calendar is used by the Coptic Church and reportedly by Egyptian
farmers.  While it normally wouldn't cross the 'official' or 'widely used'
standard for inclusion, it is computationally identical to the Ethiopian
just with a different Epoch and Month/Weekday names so is a two-for-one :-)

I'm still uncertain on the correct English transliteration for the month and
weekday names for both, as multiple versions are widely used.  Any advice on
the most common/modern forms are welcomed, but I would prefer forms that 
require only the standatd ISO-8859-1 encoding/fonts.

svn path=/trunk/KDE/kdelibs/; revision=1045676