CCMAIL: kde-buildsystem@kde.org
Nevermind, I found out how. Also, probably FindPolkitQt.cmake is another good candidate
for non-installation? It could be just copied over to kdebase/workspace/cmake/modules in 
that case. For me it's more than ok and makes even more sense, as KDE applications are not
supposed to take advantage of polkit-qt, but of KAuth instead.


svn path=/trunk/KDE/kdelibs/; revision=1051970