Quit Eventloop before emitting finished and result signals.
If KJob::exec is called then it can happen that the result(KJob*) signal is emitted before the QEventLoop in KJob::exec is quit. If a slot connected with the result(KJob*) signal does then e.g. call the same KJob::exec again then funny things may happen. The patch introduces a new internal signal that is called before the finished and result signals are emitted and that quits the eventloop. This way we can be sure that the finished and result signals are always emitted once the eventloop is done.

http://reviewboard.kde.org/r/2140/


svn path=/trunk/KDE/kdelibs/; revision=1054077