Move KDescendantsProxyModel out of stable zone.
This class uses sourceIndex.internalId() which is a major no-go in
proxy models, so it breaks very easily. It happens to work with EntityTreeModel
due to the internalIds used in that class, but breaks when used on top of a
QSortFilterProxyModel for example or a Model implemented a different way.

Something like this or a replacement will be moved back into kdelibs for 4.5.
Until then there is a non-installed copy in kdepimlibs/akonadi libkdepim.

CCMAIL: kde-pim@kde.org


svn path=/trunk/KDE/kdelibs/; revision=1054160