Revert the new cmake module committed without any review or announcement and which was in a broken state.
This has the effect that the condition for building nepomuk is now never true (instead of sometimes), but 
I have the impression that for successfully building all of KDE nepomuk is always required, since the nepomuk/trig parsing is
also used e.g. in kdepim and I think also in kdenetwork. 
So the behaviour was already wrong, and now it is more exposed. This should help getting it fixed.

Also, this commit makes explicit that soprano with redland backend and raptorparser are in fact required for building KDE since last week.

Alex

CCMAIL: trueg@kde.org
CCMAIL: kde-buildsystem@kde.org
CCMAIL: kde-core-devel@kde.org


svn path=/trunk/KDE/kdelibs/; revision=1054396