Fix reading newer XCF colormapped files
The old GIMP XCF writer had a bug with writing the PROP_COLORMAP
property. Newer versions, however, fixed that bug, and we incorrectly
assumed the bug.

The fix is to read the number of colors from the property data,
and compute the final size from that. This is how the new Krita XCF
loader does it, too.

BUG: 214384

svn path=/trunk/KDE/kdelibs/; revision=1054913