sonnet: rewrite text segmentation algorithm
since its creation sonnet was broken with indic, asian, arabic and
a lot of other languages that didn't use english-like alphabets.
this commit removes the custom text segmentation algorithm and replaces
it with a proper unicode tr29-11 algorithm found in qtextboundaryfinder
and hopefully makes sonnet work with all languages in the world. let me know
if causes any regressions.

svn path=/trunk/KDE/kdelibs/; revision=1056377