Fix cookie parsing of short expirations.
KDateTime::fromString's %Z only works if you pass the zones argument,
which is the database of zone names to match against. If you don't
pass that argument, then the resulting KDateTime is "ClockTime"

That means the time_t calculation below will be wrong.

In GMT+1, if you got a cookie with expiration date set half an hour in
the future, kcookiejar would reject the cookie because it thought it
had expired already (half an hour in the past).

svn path=/trunk/KDE/kdelibs/; revision=1070399