juggle with some code
This splits HTTPProtocol::parseContentDisposition() into two parts: one to actually parse the contents and one to store the stuff inside the HTTPProtocol object. The latter stays in place while the former is moved together with it's helper functions into parsinghelpers.cpp.

The rationale behind all this is the possibility to now add a testcase to check if the parsing works out. First test verifies that the parsing is case insensitive.

svn path=/trunk/KDE/kdelibs/; revision=1072275