Backport of commit 1055748 to the 4.3 branch:
If items are removed from the dir model, remove them also from
KFilePreviewGenerator's list of recently changed items. This should
fix crashes that occur when a file item becomes invalid just after it
is changed, e.g., if a large file is moved to a volume which is then
unmounted immediately.

Fix will be in KDE 4.3.5.

CCBUG: 211525

svn path=/branches/KDE/4.3/kdelibs/; revision=1075313