improve the Content-Disposition parser to support parameter continuations
This comes along with the corresponding testcases. The parser for the different elements (type, parameters, values) now also checks much better for invalid entries and ignores most of the junk it finds along the way.

Looking at http://greenbytes.de/tech/tc2231/ this makes us the browser with the best continuations support. We pass all the testcases for continuation support on that page except the one that also requires encoding support. Those tests are also all part of the new unit tests (in fact they were first part of my unit test during implementation and then sent to that site *g*).

svn path=/trunk/KDE/kdelibs/; revision=1075636