Fix infinite recursion issues when scrolling KCategorizedView
I don't know this code, but it seems the view is not ready to handle
QAbstractItemView::ScrollPerItem. See approx line 1222:

    if (viewMode() == ListMode && verticalScrollMode() == ScrollPerItem) {
        //TODO: think more about this case
    } else {

The default is not specified in the QAbstractItemView docs, so it may have been
changed in Qt 4.6. I have no idea.

Todays hack is to use QAbstractItemView::ScrollPerPixel by default instead.

Please re-test the bug and I will backport if it fixes Amarok and KTorrent.
I've so far only reproduced in a test app.

CCMAIL: 213068

svn path=/trunk/KDE/kdelibs/; revision=1075987