Change how we manage the proxy styles' lifetime, in order to avoid a nasty interraction which can occur if we hand-reset the style to normal on detach: if we're detaching in response to a top-levels' deletion and setStyle(0) throws oxygen back-in, it may try to create some helper widgets for effects, which will access memory of the half-destroyed top-level.
Instead, make the proxy the child of the widget --- it's deleted very late, 
so wouldn't be accessed (famous last words, I know) --- and this avoid style 
re-set. 

Will backport for rc3 despite the trickiness since it affects 
kipi-plugins, making it high-priority.

BUG: 224018

(This might also have fixed a bug with padding customization when the 
type of a listview changes. )

svn path=/trunk/KDE/kdelibs/; revision=1081278