rework the behaviour of flowing-around-floats objects /wrt floats, to match the general evolution of CSS engines.
We used to match mostly Gecko 1.8, but the behaviour has changed a lot
in the 1.9 serie. Notably, the clearing of floats by those objects is now
allowed even in auto-width containers, and generally uses the computed
width.

WebCore has some similar logic in recent trunk though the algorithm is a
bit wasteful and not so compatible in some aspects, so I only followed part
of it.

All in all, we agree the most with Opera/MSIE 7+ now (especially
considering the table-with-align-attribute behaviour).


BUG: 224228

svn path=/trunk/KDE/kdelibs/; revision=1081435