Fix paint recursion in "three bullet" password mode
There are other issues related to "three bullet" mode, such as
not being able to move cursor or select text. But simple typing
and backspace should work without 100% CPU and hangs.

Please test if this commit fixes this bug, it will be available
with the KDE SC 4.4.0 release.

CCBUG: 180482
CCBUG: 186724
CCBUG: 211250
CCBUG: 215380
CCBUG: 224693

svn path=/trunk/KDE/kdelibs/; revision=1083166