Fix Hebrew calendar dates formatted in Hebrew language.
Backport of commit 1089657:

Fix and improve Hebrew date string handling.

Restore Hebrew numbers accidentally broken in 4.4.0

Fix roundtrip of numbers <-> strings when thousands digit required

Add lots of documentation as to what it does and why

Note in 4.4 the Hebrew number code is in the base class, in 4.5 we now have a
shared-d model which allows it to be in KCalendarSystemHebrewPrivate.

Note I couldn't get the unit tests to find "he" in 4.4, they work OK in trunk,
however I have manually confirmed it works using gui stuff.


svn path=/branches/KDE/4.4/kdelibs/; revision=1094500