redirect stdout to stderr
if we are forked from a command line tool (say, ktraderclient) which is
used in a pipe, us keeping open the inherited stdout is pretty fatal for
the usefulness of the pipe.
as we have no use for a real stdout (if any at all), just redirect it to
stderr.
we can't just close/redirect stderr, as this is where all "interesting"
output is going. so if somebody tries to use some tool's stderr in a
pipe, he'd still be out of luck.

BUG: 201714

svn path=/trunk/KDE/kdelibs/; revision=1100408