Revert r861021, which added QUrl::fromPercentEncoding() to QUrl::encodedQuery().
prettyUrl() did return a prettier version of the query string, however it was broken when
the original URL contained characters such as %A0 (should be %C2%A0 to be valid UTF-8),
which were converted to UTF-8 by fromPercentEncoding().

This ends up turning %A0 into some weird, unexpected UTF-8 character in Konqueror's
location bar as well, which would pass the wrong query string when the page was reloaded.

Test cases updated.

svn path=/trunk/KDE/kdelibs/; revision=1100959