Add KMemoryInfo, a small class to provide memory information of the current system
This small class can be used to query the current OS the information about its memory status (eg total memory, free memory, etc).
Currently provides a real implementation for Linux (using sysinfo()) and for Windows (by Patrick Spendrin, using GlobalMemoryStatusEx()); for the remaining OSes there's a null implementation which always fails.
Imcluded a small unit test for the basic work (so new OS implementations can be helped in testing) and a benchmark for the update() function.

svn path=/trunk/KDE/kdelibs/; revision=1105176