Reject invalid pseudos here.
This brings our test results on Selectors API test suite to:
CSS2.1: 98.4%: 992 passed, 16 failed
CSS3:   98.9%: 2163 passed, 25 failed

16 failures on both are due to the tests expecting foo(undefined) to behave 
differently from foo(), which is a pretty odd thing to do in ECMAScript.

The 9 CSS3-only ones have to do with nested :not, probably needs a special reject. 
(Will need to bug someone for advice on that)

We do lack NAMESPACE_ERR production, though. (Which surprisingly isn't covered, and is also annoying to 
implement)

But man, this is way easier than XPath (less than a day of work!) and likely a lot faster.
Good thing I decided to put it into the feature plan, I guess.


svn path=/trunk/KDE/kdelibs/; revision=1119920