Some easy optimizations for the selectors API; bringing the total runtime on a test set down 14.8sec -> 6.8sec.
- Remove absolutely boneheaded pass-by-value (this accounted for 
much of the gain)
- Do the usual shortcircuits on ID/classname/tagname

The speed looks very competitive after this --- most of basic selectors seems faster or equal 
to FF3.6.2, though our n-th child kinda sucks. Opera 10.10 seems a bit faster, but heck, 
not bad for less than a couple of days worth of work; and the bottom-line is a nice 
performance bump for modern web frameworks.


svn path=/trunk/KDE/kdelibs/; revision=1120301