Add support for Era based calendar systems, add Japanese Era calendar system.
Date reading and formatting supports the standard POSIX %EC, %Ey, and %EY
fields.

The Japanese Era calendar supports the Nengo system from the Meiji onwards,
before this date AD is used.  More Nengo eras could be added, but consideration
should be made of the translation effort required.  Gannen (year 1 of era) is
fully supported.

For all calendars it is possible to define alternative Eras to override the 
default Eras in the global config.  It is not initially planned to support
setting these in the KCM but this option is left open.  Instead users can
manually edit the config file if really needed.

For Gregorian & Julian calendars, the AD/BC eras can be replaced with CE/BCE
via a global flag setting.  Support for setting this will be added to the KCM.

Includes a full set of unit tests.

Also includes general bugfixes uncovered by the testing :-)


svn path=/trunk/KDE/kdelibs/; revision=1121617