Add support for Thai Calendar System.
The Thai Solar Calendar System is based on the Gregorian calendar system but
using the Buddhist Era as the Epoch.  In effect, the calendar functions exactly
the same but with the year number being the Gregorian year plus 543 years.

Note that this is not an historically accurate calendar, it only implements the
scheme as used since 1941 and applies this to all days prior to this, i.e. it
does not implement the calculations used prior to this.

CCMAIL: donga.nb@gmail.com


svn path=/trunk/KDE/kdelibs/; revision=1124324