Cache when KIconLoader::loadIcon misses finding an icon.
My initial port of KIconLoader to KSharedDataCache didn't include the old behavior of
caching icon search misses. Searching for an icon and not finding it is slow! So, cache
the results.

I couldn't really understand what loadIcon was supposed to be doing so I spent some time
refactoring it instead. The behavior should be compatible with sane usage patterns, but my
first attempt didn't size "User" icons appropriately so look at this patch with
trepidation if your icons don't work.

svn path=/trunk/KDE/kdelibs/; revision=1129007