merge SVN commit 1097892 by tmcguire from 3.5 branch:
Fix regression: Use fromLatin1() instead of fromUtf8(), since we use latin1() above.
This worked previously, because local8bit() was used, which happened to be utf-8, but
got broken in r758100.

Of course the whole thing is horrible broken anyway, since it is limited to latin1 characters
and converts a QString to QCString and back, but at least this improves things a bit.

kolab/issue4054


svn path=/branches/kdepim/enterprise/kdelibs/; revision=1132145