External DocBook, step 1: remove embedded XML DTDs
This commit removes the embedded copy of the currently used XML DTDs (4.2),
and uses the external DTD instead.
If you compile and update your existing installation, you can safely remove
the existing $KDEDIR/share/kde4/apps/ksgmltools2/docbook/xml-dtd-4.2.

The code does not use catalog.xml anymore, as our custom DTD (kdex.dtd)
includes (or seems to include) all the needed files directly or indirectly.
As kdex.dtd is generated dynamically (it depends on the path to 
system-installed DTD), in order to generate the documentation during the
bootstrap phase (kdelibs documentation) the content of customization/
needs to be copied in the builddir and this copy (from the builddir itself)
is used for the generation.


svn path=/trunk/KDE/kdelibs/; revision=1134669