Don't bail out early from cmake, instead generate nice errors.
Both Strigi and DBusMenu-Qt already used macro_log_feature to create
a nice error message if they're not found. The way they use that macro
also makes sure that cmake is aborted at the point where the log is
generated so users instantly see that somethings wrong.

That makes the REQUIRED keyword for find_package less useful as in
that case cmake bails out very early leaving no way but first fixing
this error, just to eventually step into the next missing dependency
on subsequent cmake runs.

As all other dependencies in this file already use the "no REQUIRED 
keyword" but indicate in the macro_log_feature call they're required
lets do that too for Strigi and DBusMenu-Qt.

Thanks to Ingo Kloecker for pulling my attention to this.

svn path=/trunk/KDE/kdelibs/; revision=1146329