Forward port KSharedDataCache Mac OS X fixes and general end-of-world safety fixes to trunk.
This was implemented in 4.5 by Till Adam (thanks!) who noticed that it was possible to fail
to "cautious lock" the cache using CacheLocker in some situations, but the rest of the code
would assume the lock succeeded. We can't make that assumption.

He noticed this since I screwed up pthread_mutex_lock return value checking on platforms
(Mac OS X) that don't support pthread_mutex_timedlock.

svn path=/trunk/KDE/kdelibs/; revision=1148206