Formard port r1147308 from 4.5
Make KDateTime a QMetaType to allow it to be used in QVariant.  This will make
a number of fixes easier as we will be able to preserve the full KDateTime and
time spec when passing around in QVariants, rather than converting to/from 
QDateTime which is easy to get wrong.


svn path=/trunk/KDE/kdelibs/; revision=1148485