Attempt to fix a couple of KSharedDataCache bugs.
1. Stefan Majewsky has found a crasher under heavy cache load (bug 243573). The
testcase he provided did not crash for me, but did reveal that it was possible
to attempt to free a negative amount of pages, since we'd end up in the "not
enough room" case and then try to free extra room. It was possible that the way
the extra space was added caused more pages to be requested freed than were
allocated. AFAICS this should not by itself cause a crash, as removeUsedPages
notes that condition and fails. But it's still a bug. Now the number of pages
requested is capped to page table size before figuring out the number of pages
that are in use.

2. Parker Coates found a crash bug involving a cache sized just large enough to
hold exactly 1 page (the page size was == cache size). The number of entries
the cache supports is page table size / 2. 1 / 2 in integer math is 0,
therefore the entry table size was 0, and the cache eventually crashed when
trying to divide by the possible number of entries. This is fixed by ensuring
that the cache is sized to support at least a minimum number of different
pages. This means that the expectedItemSize parameter becomes more important in
determining overall minimum cache size. With default settings (4KiB page size,
256 pages) the minimum cache size is 1MiB. I also added some debugging error
messages to try and more easily diagnose these kind of logic errors in the
future.

This commit is for KDE Platform 4.6. I intend to backport to 4.5.1 as well.
CCBUG:243573

svn path=/trunk/KDE/kdelibs/; revision=1167620