Make sure that SSL metadata always gets from the ioslave all the way to the KIO::Job in the application. The main problems were: - Redirections - The slave-on-hold mechanism to hand over slaves from app (well, KRun) to app.
Redirections are solved by caching SSL metadata in the application-side
SlaveInterface object and sending it to any new job connecting to the slave.

Slave-on-hold is solved by caching the information in TCPSlaveBase and sending
it when the ioslave reconnects to the new application. A hook was added for
that.

What may or may not have aggravated the previous broken state was that the
"ssl_in_use" metadata was not stored with other ssl_* metadata in TCPSlaveBase.

BUG: 220243

svn path=/trunk/KDE/kdelibs/; revision=1170756