Backport major optimization of KIconLoader to KDE Platform 4.5.2.
This optimization is a backport of r1174319, and causes items that are
retrieved from the process-shared data cache (KSharedDataCache) to also be
stored in KIconLoader's process-local pixmap cache (QCache<>) to save the expense
of reading in the PNG data if the icon is requested in the same process in the
future.

This is backported because of the large speedup, and because this was always the
intended effect, after (silent) approval from kde-core-devel.

svn path=/branches/KDE/4.5/kdelibs/; revision=1175076