Add another "firstPage" check for KSharedDataCache.
Manuel Mommertz noticed a usage of the "firstPage" member of the KSharedDataCache
index entry (which can be set <0 to indicate it's not pointing at a page) that could
occur in sections of the code assuming the index entry was valid.

I audited the code afterwards and found another usage of firstPage without explicitly
checking to see if it could be <0, in the removeEntry function.

It turns out that every call to removeEntry had already verified that the entry
should be present. But, this makes me feel safer having the check be explicit.

svn path=/trunk/KDE/kdelibs/; revision=1175409