Correct the timeout passed to pthread_mutex_timedlock in KSharedDataCache.
Alberto Villa of the FreeBSD project noted while investigating an alternative
locking strategy for KSDC on FreeBSD that the timeout for pthread_mutex_timedlock
is an absolute point in time, not the relative 10 seconds I had been passing.

This has been working since pthread_mutex_timedlock is not required to check the
timeout if the lock can be taken, but it's still wrong.

Thanks to Alberto for the report, I will also backport to KDE Platform 4.5.2.

CCMAIL:avilla@freebsd.org

svn path=/trunk/KDE/kdelibs/; revision=1176223