Fix crash when no icon group and size is specified
Re-add the kWarning() from 4.4 branch and fall back to
KIconLoader::Desktop size in this case.

The bug in kvpnc is in KVpnc::initAction(), first three lines:
The loadIcon() calls specify no group and size. One of them
must be passed so that the loader knows which size the result
pixmap should have. It is probably simplest to replace those
calls with KIcon("iconname") constructs.

CCBUG: 246016

svn path=/trunk/KDE/kdelibs/; revision=1180089