Do not index past the very last page in the cache while defragmenting. Noted during a code review while bored in class (mental note, don't print kshareddatacache.cpp single-sided in the future).
Also, while testing an upcoming patch by Alberto Villa I had the opportunity to
test the version-change-detection code. Worked fine going from 1 to >1, but
when reverting back to 1 the cacheSize came back as 0 somehow, which caused a
hang. So check for the version being wrong but not 0 now.

This has been tested with Stefan Majewsky's KGameRenderer testbed, and might even
fix a crash in SharedMemory::defragment. I don't see how this error would cause the
Valgrind output Stefan noted, but it might indirectly cause a later operation to
fail I suppose.

This commit applies to KDE Platform 4.6. I will also backport to 4.5.2.

CCBUG:243573

svn path=/trunk/KDE/kdelibs/; revision=1180814