Backport 1181321 from trunk:
Fix painting bug on Mac OS X (though it in theory effects all platforms): sizeHint() for file item delegates was too small by one pixel because a qreal value got rounded off instead of up. So instead of "I'm a folder" the displayed file name in e.g. Dolphin was simply "I'm a fo...". Note that this was only the case for the "Short View".

svn path=/branches/KDE/4.5/kdelibs/; revision=1181329