Add support for POSIX semaphores to KSharedDataCache.
This commit is provided by Alberto Villa of the FreeBSD project, and adds support for
POSIX semaphores to the existing pthread mutex support, switchable at compile time
and at run time (by run time I mean prior to cache creation -- it's too late
afterward).

This unfortunately adds a conditional branch to every lock and unlock but that can
be eliminated with a bit of code refactoring, which I intend to undertake tonight
or tomorrow.

The code tests acceptably (both with pthreads and with semaphores forced on)
under a kind of "torture test" I have devised for the cache, basically
consisting of 4 processes hitting the same cache as fast as possible with
random data and keys.  I left it for about 15 minutes and recorded no internal
consistency errors or memory access violations. Here's hoping that is true
elsewhere as well. ;)

CCMAIL:avilla@freebsd.org

svn path=/trunk/KDE/kdelibs/; revision=1181777